update some deps

+++ src/allocator.rs
@@ -9,7 +9,7 @@ static ALLOCATOR: LockedHeap = LockedHeap::empty();
 
 fn init_heap_mapping(mapper: &mut impl Mapper<Size4KiB>, frame_allocator : &mut impl FrameAllocator<Size4KiB>) -> Result<(), MapToError<Size4KiB>>{
     let heap_start = VirtAddr::new(HEAP_START as u64);
-    let heap_end = heap_start + HEAP_SIZE - 1u64;
+    let heap_end = heap_start + (HEAP_SIZE as u64) - 1;
     let heap_start_page = Page::containing_address(heap_start);
     let heap_end_page = Page::containing_address(heap_end);
     let page_range = Page::range_inclusive(heap_start_page, heap_end_page);
